# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Core Topology Generator Web API
  version: "2026-03-05"
  description: >-
    OpenAPI specification for the CORE topology generator web backend. The API
    combines JSON endpoints used by the web UI with a set of HTML form
    submissions and download helpers. Most routes require an authenticated
    session established via `/login` and maintained with the Flask session
    cookie.
servers:
  - url: http://localhost:9090
    description: Default development server
  - url: https://localhost
    description: Reverse-proxied TLS endpoint (see README for nginx helper)
...
security:
  - sessionCookie: []
paths:
...
  /docker/cleanup:
    post:
      tags:
        - Docker
      summary: Stop and remove Docker containers created by the generator
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DockerCleanupRequest'
      responses:
        '200':
          description: Cleanup results per container.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockerCleanupResponse'
        '500':
          description: Docker command execution failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockerCleanupResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie issued after authenticating with `/login`.
  schemas:
...
    DockerCleanupRequest:
      type: object
      properties:
        names:
          type: array
          items:
            type: string
          description: Optional explicit container names to stop and remove.
    DockerCleanupResult:
      type: object
      properties:
        name:
          type: string
        stopped:
          type: boolean
        removed:
          type: boolean
    DockerCleanupResponse:
      type: object
      properties:
        ok:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/DockerCleanupResult'
        error:
          type: string
      required:
        - ok
...